+++ test/15-ports.spec.ts
@@ -1,4 +1,4 @@
-import { PortMap, PortRange } from '../src/compose/ports';
+import { PortMap, PortRange } from '~/src/compose/ports';
 import { expect } from 'chai';
 
 // Force cast `PortMap` as a public version so we can test it
@@ -298,8 +298,8 @@ describe('Ports', function () {
 
 	describe('Running container comparison', () =>
 		it('should not consider order when comparing current and target state', function () {
-			const portBindings = require('./data/ports/not-ascending/port-bindings.json');
-			const compose = require('./data/ports/not-ascending/compose.json');
+			const portBindings = require('~/test-data/ports/not-ascending/port-bindings.json');
+			const compose = require('~/test-data/ports/not-ascending/compose.json');
 			const portMapsCurrent = PortMap.fromDockerOpts(portBindings);
 			const portMapsTarget = PortMap.fromComposePorts(compose.ports);
