+++ test/27-extlinux-config.spec.ts
@@ -3,8 +3,8 @@ import { stripIndent } from 'common-tags';
 import { SinonStub, stub } from 'sinon';
 import { expect } from 'chai';
 
-import * as fsUtils from '../src/lib/fs-utils';
-import { Extlinux } from '../src/config/backends/extlinux';
+import * as fsUtils from '~/lib/fs-utils';
+import { Extlinux } from '~/src/config/backends/extlinux';
 
 describe('Extlinux Configuration', () => {
 	const backend = new Extlinux();
