+++ test/30-fdt-directive.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { FDTDirective } from '../src/config/backends/extlinux-file';
+import { FDTDirective } from '~/src/config/backends/extlinux-file';
 
 describe('FDT directive', () => {
 	const directive = new FDTDirective();
