+++ test/41-device-api-v1.spec.ts
@@ -13,29 +13,29 @@ import * as supertest from 'supertest';
 import * as path from 'path';
 import { promises as fs } from 'fs';
 
-import { exists, unlinkAll } from '../src/lib/fs-utils';
-import * as appMock from './lib/application-state-mock';
-import * as mockedDockerode from './lib/mocked-dockerode';
-import mockedAPI = require('./lib/mocked-device-api');
-import sampleResponses = require('./data/device-api-responses.json');
-import * as config from '../src/config';
-import * as logger from '../src/logger';
-import SupervisorAPI from '../src/supervisor-api';
-import * as apiBinder from '../src/api-binder';
-import * as deviceState from '../src/device-state';
-import * as apiKeys from '../src/lib/api-keys';
-import * as dbus from '../src/lib/dbus';
-import * as updateLock from '../src/lib/update-lock';
-import * as TargetState from '../src/device-state/target-state';
-import * as targetStateCache from '../src/device-state/target-state-cache';
-import blink = require('../src/lib/blink');
-import constants = require('../src/lib/constants');
-import * as deviceAPI from '../src/device-api/common';
-import { UpdatesLockedError } from '../src/lib/errors';
-import { SchemaTypeKey } from '../src/config/schema-type';
-import log from '../src/lib/supervisor-console';
-import * as applicationManager from '../src/compose/application-manager';
-import App from '../src/compose/app';
+import { exists, unlinkAll } from '~/lib/fs-utils';
+import * as appMock from '~/test-lib/application-state-mock';
+import * as mockedDockerode from '~/test-lib/mocked-dockerode';
+import mockedAPI = require('~/test-lib/mocked-device-api');
+import sampleResponses = require('~/test-data/device-api-responses.json');
+import * as config from '~/src/config';
+import * as logger from '~/src/logger';
+import SupervisorAPI from '~/src/supervisor-api';
+import * as apiBinder from '~/src/api-binder';
+import * as deviceState from '~/src/device-state';
+import * as apiKeys from '~/lib/api-keys';
+import * as dbus from '~/lib/dbus';
+import * as updateLock from '~/lib/update-lock';
+import * as TargetState from '~/src/device-state/target-state';
+import * as targetStateCache from '~/src/device-state/target-state-cache';
+import blink = require('~/lib/blink');
+import constants = require('~/lib/constants');
+import * as deviceAPI from '~/src/device-api/common';
+import { UpdatesLockedError } from '~/lib/errors';
+import { SchemaTypeKey } from '~/src/config/schema-type';
+import log from '~/lib/supervisor-console';
+import * as applicationManager from '~/src/compose/application-manager';
+import App from '~/src/compose/app';
 
 describe('SupervisorAPI [V1 Endpoints]', () => {
 	let api: SupervisorAPI;
